账本父类型选择

+++ SuiShouJi/Module/BookKeeping/Books/Model/SSJBooksTypeItem.h
@@ -26,7 +26,7 @@
 @property(nonatomic) int booksOrder;
 
 //账本父类
-@property(nonatomic) int booksParent;
+@property(nonatomic) NSInteger booksParent;
 
 @property(nonatomic, strong) NSString *userId;
 
@@ -39,4 +39,5 @@
 @property(nonatomic) BOOL editeModel;
 
 + (NSDictionary *)propertyMapping;
+
 @end
